+++ sdk/storage/storage-file-datalake/test/utils/index.ts
@@ -4,6 +4,7 @@ import { randomBytes } from "crypto";
 import * as dotenv from "dotenv";
 import * as fs from "fs";
 import * as path from "path";
+import { DefaultAzureCredential } from "@azure/identity";
 
 import { StorageSharedKeyCredential } from "../../src/credentials/StorageSharedKeyCredential";
 import { DataLakeServiceClient } from "../../src/DataLakeServiceClient";
@@ -96,6 +97,25 @@ export function getDataLakeServiceClient(
   pipelineOptions: StoragePipelineOptions = {}
 ): DataLakeServiceClient {
   return getGenericDataLakeServiceClient("DFS_", undefined, pipelineOptions);
+}
+
+export function getDataLakeServiceClientWithDefaultCredential(
+  accountType: string = "DFS_",
+  pipelineOptions: StoragePipelineOptions = {},
+  accountNameSuffix: string = ""
+): DataLakeServiceClient {
+  const accountNameEnvVar = `${accountType}ACCOUNT_NAME`;
+  let accountName = process.env[accountNameEnvVar];
+  if (!accountName || accountName === "") {
+    throw new Error(`${accountNameEnvVar} environment variables not specified.`);
+  }
+
+  const credential = new DefaultAzureCredential();
+  const pipeline = newPipeline(credential, {
+    ...pipelineOptions
+  });
+  const dfsPrimaryURL = `https://${accountName}${accountNameSuffix}.dfs.core.windows.net/`;
+  return new DataLakeServiceClient(dfsPrimaryURL, pipeline);
 }
 
 export function getAlternateDataLakeServiceClient(): DataLakeServiceClient {