+++ sdk/storage/storage-file-datalake/test/utils/testutils.common.ts
@@ -20,7 +20,10 @@ export const recorderEnvSetup: RecorderEnvironmentSetup = {
     DFS_STORAGE_CONNECTION_STRING: `DefaultEndpointsProtocol=https;AccountName=${mockAccountName};AccountKey=${mockAccountKey};EndpointSuffix=core.windows.net`,
     // Comment following line to skip user delegation key/SAS related cases in record and play
     // which depends on this environment variable
-    DFS_ACCOUNT_TOKEN: `${mockAccountKey}`
+    DFS_ACCOUNT_TOKEN: `${mockAccountKey}`,
+    AZURE_CLIENT_ID: `${mockAccountKey}`,
+    AZURE_TENANT_ID: `${mockAccountKey}`,
+    AZURE_CLIENT_SECRET: `${mockAccountKey}`
   },
   customizationsOnRecordings: [
     // Used in record mode